modified scripts for visualization, visualization_presentation is the script used to create the animations for the presentation

+++ generation/visualization_presentation.py
@@ -10,11 +10,26 @@ import cv2
 import sys
 from collections import OrderedDict
 from typing import List, Tuple, Union
-
-# Add the parent directory to the Python path
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
+from tqdm import tqdm
+from PIL import Image
+import glob
 from networks.mlp_models import MLP3D
+
+model_config = {
+        "out_size": 1,
+        "hidden_neurons": [16, 16],
+        "use_leaky_relu": False,
+        "output_type": "logits",
+        "input_dims": 2,
+        "multires": 4,
+    }
+
+def ensure_folder_exists(folder_path: str):
+    if not os.path.exists(folder_path):
+        os.makedirs(folder_path)
+        print(f"Folder '{folder_path}' created.")
+    else:
+        print(f"Folder '{folder_path}' already exists.")
 
 def make_coordinates(
     shape: Union[Tuple[int], List[int]],
@@ -54,13 +69,11 @@ def state_dict_to_min_max(state_dict: OrderedDict):
     return vmin, vmax
 
 
-def get_vmin_vmax(image_idx: int, num_epochs: int, model_config: dict):
+def get_vmin_vmax(image_idx: int, num_epochs: int, foldername: str):
     vmins = []
     vmaxs = []
-    for epoch in range(num_epochs):
-        path = os.path.dirname(os.path.abspath(__file__))
-        
-        model_path = path + f"/mnist-nerfs/recording/mnist-nerfs-unstructured-{image_idx}_{epoch}_model_final.pth"
+    for epoch in range(num_epochs):        
+        model_path = "{}/image-{}".format(foldername, image_idx) + f"_model_epoch_{epoch}.pth"
         assert os.path.exists(model_path), f"File {model_path} does not exist"
 
         vmin, vmax = state_dict_to_min_max(torch.load(model_path))
@@ -71,6 +84,20 @@ def get_vmin_vmax(image_idx: int, num_epochs: int, model_config: dict):
     vmin = torch.Tensor([v.min() for v in vmins]).min()
 
     return vmin, vmax
+
+def save_colorbar(save_path: str, vmin: torch.Tensor, vmax: torch.Tensor):
+    fig, ax = plt.subplots(figsize=(6, 1))
+    fig.subplots_adjust(bottom=0.5)
+
+    norm = plt.Normalize(vmin=vmin, vmax=vmax)
+    cb1 = plt.colorbar(
+        plt.cm.ScalarMappable(norm=norm, cmap='viridis'),
+        cax=ax, orientation='horizontal'
+    )
+    
+    cb1.set_label('Colorbar')
+    plt.savefig(save_path, transparent=True, bbox_inches='tight', pad_inches=0)
+    plt.close()
 
 def get_figure(model: torch.nn.Module, vmin: torch.Tensor, vmax: torch.Tensor):
 
@@ -128,7 +155,6 @@ def get_figure(model: torch.nn.Module, vmin: torch.Tensor, vmax: torch.Tensor):
     ax_b3.imshow(b3, cmap='viridis', aspect='equal', vmin=vmin, vmax=vmax)
     ax_b3.axis('off')
 
-
     # Save the figure
     plt.rcParams.update(
         {
@@ -136,37 +162,45 @@ def get_figure(model: torch.nn.Module, vmin: torch.Tensor, vmax: torch.Tensor):
             "savefig.facecolor": (0.0, 0.0, 0.0, 0.0),
         }
     )
-    """plt.savefig('visualization.png', dpi=300, bbox_inches='tight')
-
-    # Show the plot
-    plt.show()
-    """
 
     return fig
 
 
 
-def visualize_learning_process(image_idx: int, num_epochs: int, model_config: dict):
+def visualize_learning_process(image_idx: int, num_epochs: int, model_config: dict, foldername: str, video_name: str = "learning_process", v: tuple = None, fps: int = 10):
 
     frames_dir = "frames"
 
-    vmin, vmax = get_vmin_vmax(image_idx, num_epochs, model_config)
+    frames = []
 
+    if v:
+        vmin, vmax = v
+    else:
+        vmin, vmax = get_vmin_vmax(image_idx, num_epochs, foldername)
 
-    for epoch in range(num_epochs):
-        path = os.path.dirname(os.path.abspath(__file__))
-        model = MLP3D(**model_config)
-        
-        model_path = path + f"/mnist-nerfs/recording/mnist-nerfs-unstructured-{image_idx}_{epoch}_model_final.pth"
-        assert os.path.exists(model_path), f"File {model_path} does not exist"
+    with tqdm(total=num_epochs) as pbar:
+        for epoch in range(num_epochs):
+            model = MLP3D(**model_config)
+            
+            model_path = "{}/image-{}".format(foldername, image_idx) + f"_model_epoch_{epoch}.pth"
+            if os.path.exists(model_path) is False:
+                print(f"File {model_path} does not exist")
+                return
 
-        model.load_state_dict(torch.load(model_path))
+            model.load_state_dict(torch.load(model_path))
 
-        fig = get_figure(model, vmin, vmax)
+            fig = get_figure(model, vmin, vmax)
 
-        frame_path = os.path.join(frames_dir, f"frame_{epoch:03d}.png")
-        plt.savefig(frame_path, format="png")
-        plt.close(fig)
+            frame_path = os.path.join(frames_dir, f"frame_{epoch:03d}.png")
+            plt.savefig(frame_path, format="png")
+
+            plt.close(fig)
+
+            pbar.update(1)
+            pbar.set_description(
+                    "Finished creating frame for epoch %d"
+                    % (epoch)
+                )
 
     # Compile the images into a video
     frame_paths = [
@@ -176,33 +210,93 @@ def visualize_learning_process(image_idx: int, num_epochs: int, model_config: di
     ]
     frame = cv2.imread(frame_paths[0])
     height, width, layers = frame.shape
+
     
-    video_path = "learning_process.mp4"
-    out = cv2.VideoWriter(video_path, cv2.VideoWriter_fourcc(*"mp4v"), 10, (width, height))
+    video_path = video_name + ".mp4"
+    gif_path = video_name + ".gif"
+    out = cv2.VideoWriter(video_path, cv2.VideoWriter_fourcc(*"mp4v"), fps, (width, height))
     
     for frame_path in frame_paths:
         frame = cv2.imread(frame_path)
+        frames.append(frame)
         out.write(frame)
-
+    
     out.release()
     print(f"Video saved to {video_path}")
 
+    images = [Image.open(frame_path).convert("RGBA") for frame_path in frame_paths]
+
+    # Save as GIF
+    images[0].save(
+        gif_path,
+        save_all=True,
+        append_images=images[1:],
+        duration=1/fps,  # duration in milliseconds
+        loop=0,
+        transparency=0,  # specify the transparency color (usually 0 for black)
+        disposal=2  # ensures that each frame is replaced, not drawn on top of the previous one
+    )
+
+    print(f"GIF saved to {gif_path}")
+
 
 def main():
-    model_config = {
-        "out_size": 1,
-        "hidden_neurons": [16, 16],
-        "use_leaky_relu": False,
-        "output_type": "logits",
-        "input_dims": 2,
-        "multires": 4,
-    }
 
-    image_idx = 0  # Change this to visualize a different image from the dataset
-    num_epochs = 200  # Number of epochs or models saved
-    visualize_learning_process(image_idx, num_epochs, model_config)
+    image_idx = 1721  # Change this to visualize a different image from the dataset
+    num_epochs = 250  # Number of epochs or models saved
+    subfoldername = "unconditioned" #"pretrained"
+    foldername = (
+        f"./datasets/mnist-nerfs/{subfoldername}"
+    )
+    visualize_learning_process(image_idx, num_epochs, model_config, foldername, f"./animation/{subfoldername}/presentation_{subfoldername}_{image_idx}")
 
+def create_all_videos():
+    for filename in glob.glob("./datasets/mnist-nerfs/unconditioned/*_final.pth"):
+
+        image_idx = filename.split("_")[-3].split("-")[-1]  # Change this to visualize a different image from the dataset
+        num_epochs = 250  # Number of epochs or models saved
+        subfoldername = "unconditioned" #"pretrained"
+        foldername = (
+            f"./datasets/mnist-nerfs/{subfoldername}"
+        )
+        visualize_learning_process(image_idx, num_epochs, model_config, foldername, f"./animation/{subfoldername}/presentation_{subfoldername}_{image_idx}")
+
+    for filename in glob.glob("./datasets/mnist-nerfs/pretrained/*_final.pth"):
+
+        image_idx = filename.split("_")[-3].split("-")[-1]  # Change this to visualize a different image from the dataset
+        num_epochs = 250  # Number of epochs or models saved
+        subfoldername = "pretrained" #"pretrained"
+        foldername = (
+            f"./datasets/mnist-nerfs/{subfoldername}"
+        )
+        visualize_learning_process(image_idx, num_epochs, model_config, foldername, f"./animation/{subfoldername}/presentation_{subfoldername}_{image_idx}")
+
+def compare_different_runs(image_idxs: List[int], num_epoch: int, model_config: dict, subfoldernames: List[str]):
+    vmins = []
+    vmaxs = []
+
+    # find max and min of v
+    for idx in image_idxs:
+        for subfoldername in subfoldernames:
+            vmin, vmax = get_vmin_vmax(idx, num_epoch,  f"./datasets/mnist-nerfs/{subfoldername}")
+            vmins.append(vmin)
+            vmaxs.append(vmax)
+    vmax = torch.Tensor([v.max() for v in vmaxs]).max()
+    vmin = torch.Tensor([v.min() for v in vmins]).min()
+
+    save_folder = f"./animation/comparison_{'_'.join(map(str, image_idxs))}"
+    
+    for idx in image_idxs:
+        for subfoldername in subfoldernames:
+            foldername =  f"./datasets/mnist-nerfs/{subfoldername}"
+            video_name = save_folder + f"/{subfoldername}_{idx}"
+
+            ensure_folder_exists(save_folder)
+            visualize_learning_process(idx, num_epoch, model_config, foldername,  video_name=video_name, v= (vmin, vmax))
+    
+    save_colorbar(save_folder + "/colorbar.png", vmin, vmax)
 
 if __name__ == "__main__":
-    main()
+    compare_different_runs([0, 11, 35, 47, 65], 200, model_config, ["pretrained", "unconditioned"])
+
